requirements.txt
pytest~=9.0.3
python-dotenv~=1.2.2
requests~=2.33.1